status-api/requirements.txt
Flask==1.1.4
Jinja2==2.11.3
MarkupSafe==2.0.1
Werkzeug==1.0.1
aniso8601==9.0.1
arrow==1.2.1
attrs==21.2.0
blinker==1.4
cachelib==0.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==7.1.2
ecdsa==0.17.0
flask-jwt-oidc==0.3.0
flask-restx==0.5.1
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
itsdangerous==1.1.0
jsonschema==4.3.0
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
pytz==2021.3
requests==2.26.0
rsa==4.8
sentry-sdk==1.5.1
six==1.16.0
urllib3==1.26.7
zipp==3.6.0
-e git+https://github.com/bcgov/sbc-common-components.git#egg=sbc-common-components&subdirectory=python
